try new requirments

+++ src/confidence_score_test.py
@@ -70,7 +70,7 @@ def compute_bert_recall_confidence_score(cv_path, job_desc_path, cover_letter_pa
     similarity = cosine_similarity([combined_cv_jd], [cl_embedding])[0][0]
 
     # Return the similarity score as a percentage
-    return int(similarity * 100)
+    return float(similarity * 100)
 
 if __name__ == "__main__":
     # Ensure correct number of arguments
